+++ src/Table/Interfaces/Cells/SupportsIcon.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Patrikjak\Utils\Table\Interfaces\Cells;
 
-use Patrikjak\Utils\Common\Enums\Icon;
+use Patrikjak\Utils\Common\Icon;
 
 interface SupportsIcon
 {